bplus_layout: the unsafe primitives leave the public surface
arr_get, slice_get, arr_set, arr_shift_up and sel_usize go pub(crate):
they lower to get_unchecked under verified preconditions, so misuse
from outside the crate was the one place where 'unguarded' meant
undefined behavior rather than a panic or a wrong answer. Their
contract tests move in-crate (src/bplus_layout_tests.rs, merged from
the former integration proptest suite and the LCG fuzz block), since
external tests can no longer name them - which is the point.

Each unsafe body also gains a debug_assert on its verified
precondition: a runtime monitor for crate-internal unverified callers,
free in release, catching a violated bound as a clean panic in every
debug and fuzz run instead of an out-of-bounds access.

containers-verus 1501/0 verified; workspace tests 1313/0; trust
surface unchanged at 26/5.
