+++ containers-verus/tests/external_body_contract_fuzz.rs
@@ -344,153 +344,3 @@ fn shrink_preserves_aov_contents() {
     }
     println!("shrink_preserves_aov_contents: OK (200 rounds)");
 }
-
-// --------------------------------------------------------------------------
-// bplus_layout::arr_get / arr_set / sel_usize / arr_shift_up.
-//
-// (`slice_get`, the fifth bplus_layout trusted primitive, is covered by the
-// property suite `tests/bplus_layout_proptest.rs`, which also re-checks these
-// four with shrinking and the tree's actual `(T, N)` instantiations.)
-//
-// These four are `external_body` so that a fact Verus already PROVED reaches
-// the machine code: the elided bounds check (`arr_get`/`arr_set`, whose `i < N`
-// is a verified precondition at every call site), the `cmov` lowering of the
-// bisection's data-dependent update (`sel_usize`), and one `memmove` where
-// Verus's invariant rules force an element loop (`arr_shift_up`). Their
-// contracts are trivial to state and therefore trivial to check at runtime,
-// which is exactly why they are fuzzed rather than argued about:
-// `arr_get`/`arr_set` must agree with checked indexing on every in-range index,
-// `sel_usize` must agree with the `if`/`else` it replaces on both boolean
-// values, and `arr_shift_up` must agree with `copy_within` at every
-// `(pos, cnt)` -- including across its internal length dispatch, which is the
-// one place a hand-written fast path could disagree with the slow one.
-//
-// `arr_get`/`arr_set` are only called OUT of range by a caller that violated a
-// verified precondition, so the out-of-range case is deliberately not exercised
-// (it is UB, not a testable branch).
-// --------------------------------------------------------------------------
-
-#[test]
-fn arr_get_set_agree_with_checked_indexing() {
-    use semi_persistent_containers_verus::bplus_layout::{arr_get, arr_set};
-
-    const N: usize = 62;
-    let mut lcg: u64 = 0x5EED_1234_ABCD_0001;
-    let mut next = move || {
-        lcg = lcg
-            .wrapping_mul(6364136223846793005)
-            .wrapping_add(1442695040888963407);
-        lcg >> 1
-    };
-
-    for round in 0..500 {
-        // A fresh array and a mirror the test indexes the ordinary, checked way.
-        let mut a = [0u32; N];
-        let mut mirror = [0u32; N];
-        for (i, m) in mirror.iter_mut().enumerate() {
-            let v = next() as u32;
-            arr_set(&mut a, i, v);
-            *m = v;
-        }
-        // reads agree at every in-range index
-        for (i, &m) in mirror.iter().enumerate() {
-            assert_eq!(
-                arr_get(&a, i),
-                m,
-                "round {round}: arr_get disagrees with checked read at {i}"
-            );
-        }
-        // an overwrite touches exactly the one slot it names
-        let pos = (next() as usize) % N;
-        let w = next() as u32;
-        arr_set(&mut a, pos, w);
-        mirror[pos] = w;
-        for (i, &m) in mirror.iter().enumerate() {
-            assert_eq!(
-                arr_get(&a, i),
-                m,
-                "round {round}: arr_set perturbed slot {i} while writing {pos}"
-            );
-        }
-    }
-    println!("arr_get_set_agree_with_checked_indexing: OK (500 rounds x 62 slots)");
-}
-
-#[test]
-fn sel_usize_agrees_with_if_else() {
-    use semi_persistent_containers_verus::bplus_layout::sel_usize;
-
-    let mut lcg: u64 = 0x9E37_79B9_7F4A_7C15;
-    let mut next = move || {
-        lcg = lcg
-            .wrapping_mul(6364136223846793005)
-            .wrapping_add(1442695040888963407);
-        lcg >> 1
-    };
-
-    // Edge values plus random ones: the postcondition is `if c { b } else { a }`,
-    // so both arms must be reproduced exactly, including at the type's extremes.
-    let mut cases: Vec<(usize, usize)> = vec![
-        (0, 0),
-        (0, usize::MAX),
-        (usize::MAX, 0),
-        (usize::MAX, usize::MAX),
-        (1, usize::MAX - 1),
-    ];
-    for _ in 0..500 {
-        cases.push((next() as usize, next() as usize));
-    }
-
-    for (i, &(a, b)) in cases.iter().enumerate() {
-        for c in [false, true] {
-            let expect = if c { b } else { a };
-            assert_eq!(
-                sel_usize(c, a, b),
-                expect,
-                "case {i}: sel_usize({c}, {a}, {b}) != if/else"
-            );
-        }
-    }
-    println!(
-        "sel_usize_agrees_with_if_else: OK ({} cases x 2)",
-        cases.len()
-    );
-}
-
-#[test]
-fn arr_shift_up_agrees_with_copy_within() {
-    use semi_persistent_containers_verus::bplus_layout::arr_shift_up;
-
-    const N: usize = 62;
-    let mut lcg: u64 = 0xD1B5_4A32_D192_ED03;
-    let mut next = move || {
-        lcg = lcg
-            .wrapping_mul(6364136223846793005)
-            .wrapping_add(1442695040888963407);
-        lcg >> 11
-    };
-
-    // Every (pos, cnt) with pos <= cnt < N is exercised exhaustively, so both
-    // arms of the internal length dispatch (scalar below the measured crossover,
-    // `memmove` above it) are covered, as is the boundary between them -- the
-    // only place the optimization could introduce a disagreement.
-    let mut checked = 0usize;
-    for cnt in 0..N {
-        for pos in 0..=cnt {
-            let mut a = [0u32; N];
-            for slot in a.iter_mut() {
-                *slot = next() as u32;
-            }
-            let mut want = a;
-            want.copy_within(pos..cnt, pos + 1);
-
-            arr_shift_up(&mut a, pos, cnt);
-            assert_eq!(
-                a, want,
-                "arr_shift_up(pos={pos}, cnt={cnt}) disagrees with copy_within"
-            );
-            checked += 1;
-        }
-    }
-    println!("arr_shift_up_agrees_with_copy_within: OK ({checked} (pos, cnt) pairs)");
-}
